initial changes towards `Explorer` refactoring
